bugfixing

+++ ZPO6.2/Serwisy/DishService.cs
@@ -38,7 +38,7 @@ namespace ZPO6._2
         public List<Potrawa> SelectAllDishes()
             => dish.GetAll();
 
-        public Potrawa CreateDish(string name, int price, int possible, bool isVege)
+        public Potrawa CreateDish(string name, decimal price, int possible, bool isVege)
         {
             Potrawa model = new Potrawa();
             model.Nazwa = name;
